refactor: remove remaining legacy shims
