+++ docs/reports/v0.1.4b/high-impact-ideas.md
@@ -2,4 +2,4 @@
 
 1. **Surface simulation diagnostics in the developer HUD.** The runtime already tracks detailed Rapier panic and subsystem failure metadata on `GameState.simulation.rapierDiagnostics`, but the data is only written by `recordRapierStepPanic`/`recordSubsystemFailure` and never exposed to the UI. Adding a HUD panel that streams these counters (including the snapshots published for `copilot_debug`) would shorten the feedback loop for physics or AI regressions and make invisible failures actionable during playtesting.【F:src/game/state.ts†L72-L105】【F:src/game/simulationQueue.ts†L55-L107】
 2. **Move thruster trails to GPU-managed buffers.** The current particle trail system preloads all GLTFs, clones vectors for every spawn, and iterates a large particle pool on the CPU each frame, which will bottleneck large fleet scenarios.【F:src/components/ParticleTrails.tsx†L41-L200】 Replacing the pool with instanced buffer attributes (e.g., `InstancedBufferGeometry` with shader-driven fading) or a compute/transform feedback approach would shift the workload to the GPU and cut allocations, improving scalability.
-3. **Package the AI scenario harness as a CLI tool.** `runAIScenario` already composes deterministic seeds, logging, and metrics aggregation, but there is no npm script or entrypoint that lets designers run canned matchups outside of tests.【F:src/game/aiScenarioHarness.ts†L31-L182】【F:package.json†L6-L31】 Wrapping the harness in a `scripts/` command (e.g., `npm run sim:scenario -- --config scenario.json`) would unlock rapid balance sweeps and integrate the analytics pipeline into CI.
+3. **Package the AI scenario harness as a CLI tool.** `runAIScenario` already composes deterministic seeds, logging, and metrics aggregation, but there is no npm script or entrypoint that lets designers run canned matchups outside of tests.【F:test/support/aiScenarioHarness.ts†L31-L182】【F:package.json†L6-L31】 Wrapping the harness in a `scripts/` command (e.g., `npm run sim:scenario -- --config scenario.json`) would unlock rapid balance sweeps and integrate the analytics pipeline into CI.
