--- src/components/ExplosionRenderer.tsx
@@ -1,7 +1,0 @@
-/**
- * Legacy export for backward compatibility.
- * The implementation has been refactored into modular components.
- * @deprecated Import from './explosions/ExplosionRendererCore.js' instead.
- */
-export { ExplosionRendererCore as ExplosionRenderer, ExplosionsLayer } from './explosions/ExplosionRendererCore.js';
-
