--- src/game/aiScenarioHarness/stateBuilder.ts
@@ -1,7 +1,0 @@
-/**
- * @deprecated Use direct imports from shipFactory and stateFactory instead.
- * This file is maintained for backward compatibility.
- */
-
-export { createHarnessShip } from './shipFactory.js';
-export { createHarnessState } from './stateFactory.js';
