+++ test/support/aiScenarioHarness/rapierShim.ts
